[#74]refactor: stmt model들 모두 파일 분리
