+++ app/visualize/generator/model/expr_viz.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class ExprViz:
+    id: int
+    depth: int
+    expr: str
+    highlights: []
+    type: str
